Planets
Gravity wells around planets.

THEY DO NOT KILL YOU YET

But as you approach a planet, you will be pulled in towards it. the
closer you get, the stronger the force.

Put the script planetScript onto everything that you want to be affected
by planets, and make sure that planets have the planet tag.